Admin Tool opens within the same file
I made the main JFrame global so that SideBar.java could access it
without having the jf object from Main.java being passed to it. I could
have altered it, but it would have been a huge workaround so I made a
quick hack and made it global. Yes, I realize global variables are not
recommended, but it was the easiest way to have one JFrame without
altering the code too much.