Forgot to add campus.jpg to Maps folder and updated other files
Moved campus.jpg to MapUI/Maps and updated FileChooser.java and
AdminTool.java
Delete functionality is there, but still shows up on the application
(but is deleted from the project itself)
Adding the map doesn't dynamically refresh the combo box at the moment.
If the application is re-run, it will appear.